service broker should assign config file path when start
[#133527149]
