+++ servicebroker/test/routes/catalog_test.js
@@ -4,15 +4,16 @@ var expect = require('chai').expect;
 
 var fs = require('fs');
 var path = require('path');
+var configFilePath = path.join(__dirname, '../../config/settings.json');
 var settings = require(path.join(__dirname, '../../lib/config/setting.js'))((JSON.parse(
-  fs.readFileSync(path.join(__dirname, '../../config/settings.json'), 'utf8'))));
+  fs.readFileSync(configFilePath, 'utf8'))));
 var auth = new Buffer(settings.username + ":" + settings.password).toString('base64');
 
 describe('getCatalog RESTful API', function() {
   var server;
   beforeEach(function() {
-    delete require.cache[require.resolve('../../lib/index.js')];
-    server = require(path.join(__dirname, '../../lib/index.js'));
+    delete require.cache[require.resolve('../../lib/server.js')];
+    server = require(path.join(__dirname, '../../lib/server.js'))(configFilePath);
   });
 
   afterEach(function(done) {
