+++ servicebroker/test/routes/serviceinstance_test.js
@@ -4,8 +4,9 @@ var uuid = require('uuid');
 
 var fs = require('fs');
 var path = require('path');
+var configFilePath = path.join(__dirname, '../../config/settings.json');
 var settings = require(path.join(__dirname, '../../lib/config/setting.js'))((JSON.parse(
-  fs.readFileSync(path.join(__dirname, '../../config/settings.json'), 'utf8'))));
+  fs.readFileSync(configFilePath, 'utf8'))));
 
 var models = require('../../lib/models')(settings.db.uri);
 var serviceInstance = models.service_instance;
@@ -22,8 +23,8 @@ describe('service instance RESTful API', function() {
 
 
   before(function() {
-    delete require.cache[require.resolve('../../lib/index.js')];
-    server = require(path.join(__dirname, '../../lib/index.js'));
+    delete require.cache[require.resolve('../../lib/server.js')];
+    server = require(path.join(__dirname, '../../lib/server.js'))(configFilePath);
   });
 
   after(function(done) {
